# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación realizado con Swagger, servicios funcionando con Laravel",
        "description": "Proyecto para la comunicación REST",
        "contact": {
            "name": "Correo de contacto",
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/service/rest/v1/security/user": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Muestra todos los usuarios registrados en la base de datos.",
                "description": "Retorna un array de usuarios.",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Crear un nuevo usuario en el sistema.",
                "description": "Retorna el objeto del usuario creado.",
                "operationId": "postUsers",
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/service/rest/v1/security/user_password": {
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Cambiar la contraseña del usuario seleccionado.",
                "description": "Retorna el objeto del usuario que cambio la contraseña.",
                "operationId": "postUserPassword",
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/service/rest/v1/security/user/{user}": {
            "put": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Actualizar el usuario seleccionado.",
                "description": "Retorna el objeto del usuario actualizado.",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID del usuario para actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Eliminar o cambiar el estado del usuario seleccionado.",
                "description": "Retorna el objeto del usuario eliminado.",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID del usuario para eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
        "/service/rest/v1/security/user_rol/{user_rol} ": {
            "get": {
                "tags": [
                    "Usuario y Rol"
                ],
                "summary": "Muestra todos los roles asignados al usuario seleccionado.",
                "description": "Retorna un array de roles del usuario seleccionado.",
                "operationId": "findUserRolbyId",
                "parameters": [
                    {
                        "name": "user_rol",
                        "in": "path",
                        "description": "ID del usuario para buscar los roles",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuario y Rol"
                ],
                "summary": "Eliminar el rol asignado al usuario seleccionado.",
                "description": "Retorna el objeto del rol asignado al usuario eliminado.",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "user_rol",
                        "in": "path",
                        "description": "ID del rol para eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}